Refactored unit tests related to bucket map into test_config_util.py
